ludo_kill_websocket/requirements.txt
Django==2.1.3
djangorestframework==3.9.0
channels==2.1.5
websockets==7.0
channels-redis==2.3.1